# api.yaml
openapi: 3.0.1
info:
  title: NDO Slackbot POC
  description: NDO Slackbot API
  version: 0.3.0
servers:
  - url: http://localhost:3000/
    description: Local development URL
paths:
  "/slack/plainCard":
    post:
      tags:
        - Slack Notification API
      summary: Post plain message
      description: Posts a plain message to the channel corresponding to the webhook
      operationId: plainCardSlackMessage
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SlackNotificationRequest"
        required: true
      responses:
        '200':
          description: Successful post
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '500':
          description: Service error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
  "/slack/hello":
    post:
      tags:
        - Slack Notification API
      summary: Say Hello
      description: Posts a greeting for a particular user to the channel corresponding
        to the webhook
      operationId: helloSlack
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SlackNotificationRequest"
        required: true
      responses:
        '200':
          description: Successful post
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '500':
          description: Service error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
  "/slack/alert":
    post:
      tags:
        - Slack Notification API
      summary: Post alert message
      description: Posts an alert message to the channel corresponding to the webhook
      operationId: postAlertMessage
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AlertRequest"
        required: true
      responses:
        '200':
          description: Alert successfully posted
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
        '500':
          description: Service error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Response"
components:
  schemas:
    AlertRequest:
      allOf:
        - $ref: '#/components/schemas/SlackNotificationRequest'
        - type: object
          required:
            - title
            - level
            - webhookUrl
          properties:
            level:
              type: string
              enum:
                - ERROR
                - WARN
                - OPERATIONAL
            title:
              type: string
    SlackNotificationRequest:
      type: object
      required:
        - webhookUrl
      properties:
        content:
          type: string
        webhookUrl:
          type: string
    Response:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        contentType:
          type: string
        headers:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        body:
          type: string